requirements.txt
numpy
scipy
tqdm
lpips
mmcv
imageio
imageio-ffmpeg
opencv-python
pytorch_msssim
torch
torch_scatter
Pillow